requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
discord.py==2.3.2
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
ipapi==1.0.4
kaitaistruct==0.10
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.1
psutil==5.9.8
pyasn1==0.5.1
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
selenium==4.18.1
selenium-wire==5.1.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
vboxapi==1.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
zstandard==0.22.0
